// nodes/CodeKit/actions/pdf/operation.types.ts
export enum OperationType {
	COMPRESS = 'compress',
	COUNT = 'count',
	CREATE = 'create',
	DECRYPT = 'decrypt',
	DOCX_TO_PDF = 'docx-to-pdf',
	DRAW_IMAGE = 'draw-image',
	DRAW_TEXT = 'draw-text',
	ENCRYPT = 'encrypt',
	HTML = 'html',
	MARKDOWN_TO_PDF = 'markdownstringtopdf',
	METADATA = 'metadata',
	MERGE = 'merge',
	PAGES_ADD = 'pages-add',
	PAGES_REMOVE = 'pages-remove',
	PAGES_RESIZE = 'pages-resize',
	PAGES_ROTATE = 'pages-rotate',
	BASE64 = 'base64',
	PDF_TO_IMAGE = 'pdf-to-image',
	SPLIT = 'split',
	WATERMARK_TEXT = 'watermark-text',
	WATERMARK_IMAGE = 'watermark-image',
}